# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bütçe API",
        "version": "1.0"
    },
    "paths": {
...
        "/workplaces/findAll": {
            "get": {
                "tags": [
                    "Workplaces"
                ],
                "summary": "Find all workplaces",
                "operationId": "workplaces_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkplacesFindAllResultVM"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/workplaces/findAllByPaging": {
            "get": {
                "tags": [
                    "Workplaces"
                ],
                "summary": "Find all workplaces by pagination",
                "operationId": "workplaces_findAllByPaging",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_perPage",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_sort",
                        "in": "query",
                        "description": "Order items",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkplaceFindAllByPageResultVM"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/workplaces/findOneById/{workplaceId}": {
            "get": {
                "tags": [
                    "Workplaces"
                ],
                "summary": "Find workplace data",
                "operationId": "workplaces_findOneById",
                "parameters": [
                    {
                        "name": "workplaceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkplaceFindOneByIdResultVM"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/workplaces/create": {
            "post": {
                "tags": [
                    "Workplaces"
                ],
                "summary": "Create a new workplace",
                "operationId": "workplaces_create",
                "requestBody": {
                    "description": "Create new workplace",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkplaceSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/workplaces/update/{workplaceId}": {
            "put": {
                "tags": [
                    "Workplaces"
                ],
                "summary": "Update Workplace",
                "operationId": "workplaces_update",
                "parameters": [
                    {
                        "name": "workplaceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Workplace",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkplaceSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/workplaces/delete/{workplaceId}": {
            "delete": {
                "tags": [
                    "Workplaces"
                ],
                "summary": "Delete workplace",
                "operationId": "workplaces_delete",
                "parameters": [
                    {
                        "name": "workplaceId",
                        "in": "path",
                        "description": "Workplace uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/years/findAll": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Find all years",
                "operationId": "years_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/YearsFindAllResultVM"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ObjectId": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "WorkplaceFindAllByPageResultVM": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/WorkplaceListItem"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "WorkplaceFindOneById": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "companyId": {
                        "type": "string"
                    },
                    "companyName": {
                        "type": "string"
                    },
                    "workplaceName": {
                        "type": "string"
                    },
                    "registrationNumber": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "WorkplaceFindOneByIdResultVM": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/WorkplaceFindOneById"
                    }
                },
                "type": "object"
            },
...
            "WorkplaceListItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "companyName": {
                        "type": "string"
                    },
                    "workplaceName": {
                        "type": "string"
                    },
                    "registrationNumber": {
                        "description": "maxLength: 100",
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "WorkplaceSave": {
                "properties": {
                    "workplaceId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "companyId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "workplaceName": {
                        "description": "maxLength: 255",
                        "type": "string"
                    },
                    "registrationNumber": {
                        "description": "maxLength: 100",
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "WorkplacesFindAllResultVM": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "YearsFindAllResultVM": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}